Adding the file 5, updating the header and fixing file 3 so index receives unsigned long int values

+++ 0x1A-hash_tables/3-hash_table_set.c
@@ -12,7 +12,7 @@
 int hash_table_set(hash_table_t *ht, const char *key, const char *value)
 {
 	hash_node_t *list_n = NULL, *new = NULL;
-	int idx = 0;
+	unsigned long int idx = 0;
 
 	if (ht == NULL)
 		return (0);
